crawling_test/requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
supabase>=2.1.0
openai>=1.3.0
python-dotenv>=1.0.0
httpx[http2]>=0.26,<0.29
